Renamed inbox operations to be more consistent
